partial proofs for additive identity
I started the process of proving that the sum of a rational number and it's inverse is 0. For this, I needed the type ZZ (which is an implementation of that integers that allows proofs). I added a few functions for converting between pairs of ZZ and integers. I believe that using cong, the next step (showing that x = -y implies that x+y = 0) can be used to show that the first number in the pair of addition of a number and its inverse is always zero.

A similar process can be used for multiplicative inverses, but will involve simplification of the rational generated and showing that it is equal to (1,1).

Documentation for ZZ:

https://github.com/idris-lang/Idris-dev/blob/master/libs/contrib/Data/ZZ.idr