Rationals which are Integers
The function `IsRationalZ` checks with proof whether a rational number is equal to an integer. If yes, it gives the integer which it is equal to; otherwise, it shows that the remainder on division of the numerator by the denominator is nonzero. The function `nonZeroNotZero` in ZZUtils has holes to be filled in though.

-- Abishek (AR-MA210)