simplification with proof + equality of rationals
I added a function which takes a rational number and produces the simplified form with proof.
Also, I started working on a custom equality type for rationals (a/b = c/d if ad=bc). I proved reflexivity and symmetry for this type. I am trying to be as careful as possible with this equality type.

-- Abishek (AR-MA210).